chore: Resolving conflicts;
chore: add migration and seeding for systems table (#1246)
