+++ client/src/api/schemas/codeSystemsReponse.ts
@@ -1,0 +1,10 @@
+
+/**
+ * Display information needed for code system information on the frontend.
+ */
+export interface CodeSystemsReponse {
+  id: string;
+  key: string;
+  display_name: string;
+  oid: string;
+}
