+++ client/src/api/schemas/index.ts
@@ -10,6 +10,7 @@ export * from './bodyUploadEcr';
 export * from './codeCategoryCompletenessStatus';
 export * from './codeSetStatus';
 export * from './codeSystem';
+export * from './codeSystemsReponse';
 export * from './completenessStatus';
 export * from './condition';
 export * from './configurationCustomCodeResponse';
