+++ refiner/app/api/v1/v1_router.py
@@ -1,6 +1,7 @@
 from fastapi import APIRouter
 
 from .app_notifications import router as app_notifications_router
+from .code_systems import router as code_systems_router
 from .conditions import router as conditions_router
 from .configurations.router import router as configurations_router
 from .demo import router as demo_router
@@ -15,4 +16,5 @@ router.include_router(demo_router)
 router.include_router(events_router)
 router.include_router(info_router)
 router.include_router(releases_router)
+router.include_router(code_systems_router)
 router.include_router(app_notifications_router)
